move pkg/client functions to pkg/utils
