+++ pkg/utils/kubernetes/deployment_test.go
@@ -14,7 +14,7 @@ import (
 	appsv1 "k8s.io/api/apps/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	"github.com/gardener/gardener/pkg/client/kubernetes"
+	"github.com/gardener/gardener/pkg/utils/kubernetes"
 	"github.com/gardener/gardener/pkg/utils/retry"
 	mockclient "github.com/gardener/gardener/third_party/mock/controller-runtime/client"
 )
